+++ lib/src/main/java/com/stepango/rxdatabindings/ObservableString.kt
@@ -31,7 +31,8 @@ class ObservableString(value: String = "") : ObservableField<String>(value), Par
     companion object {
         internal const val serialVersionUID = 1L
 
-        @JvmStatic val CREATOR: Parcelable.Creator<ObservableString> = object : Parcelable.Creator<ObservableString> {
+        @JvmField
+        val CREATOR: Parcelable.Creator<ObservableString> = object : Parcelable.Creator<ObservableString> {
 
             override fun createFromParcel(source: Parcel) = ObservableString(source.readString())